ca y est, la règle s'applique et le 1er test passe

+++ src/main/kotlin/location/domain/usine/regles/CalculPrixHauteSaison.kt
@@ -4,8 +4,9 @@ import location.domain.usine.PourCalculerLePrix
 import location.domain.valueObjects.Monnaie
 
 class CalculPrixHauteSaison : PourCalculerLePrix {
+
     override fun calculPrix(duree: Int): Monnaie {
 
-      return  Monnaie.Euros(  duree / 60)
+      return  Monnaie.Euros(  duree.toDouble() / 60.0 )
     }
 }
